Migrate class constructors into their own module
