+++ src/logic.js
@@ -1,33 +1,5 @@
-class Project {
-
-  constructor(title) {
-    this.title = title;
-    this.todos = [];
-  }
-
-  addTodo = (todo) => {
-    todos.push(todo);
-  }
-
-  getTodo = (todoTitle) => {
-    const myTodo;
-    todos.forEach((todo) => {
-      if (todo.getTitle() === todoTitle) {
-        myTodo = todo;
-      }
-    });
-    return myTodo;
-  }
-
-  deleteTodo = (todoTitle) => {
-    todos.forEach((todo, idx) => {
-      if (todo.getTitle() === todoTitle) {
-        todos.splice(idx, 1);
-      }
-    });
-  }
-}
-
+import Project from './classes.js'
+import Todo from './classes.js'
 const allProjects = [];
 
 // Loop thru each project and display it.
@@ -98,23 +70,4 @@ const createTodo = () => {
 
   })
 };
-
-
-
-
-
-class Todo {
-
-  constructor(title, description, priority, dueDate) {
-    this.title = title;
-    this.description = description;
-    this.priority = priority;
-    this.dueDate = dueDate;
-  }
-
-  assignPriority = (givenPriority) => {
-    priority = givenPriority
-  }
-}
-
